miniapp/backend/requirements.txt
python-telegram-bot==20.7
requests==2.32.3
Flask==3.0.3
fastapi==0.115.0
uvicorn==0.30.1
requests==2.32.3
gspread==6.1.2
pydantic==2.9.2
google-api-python-client==2.115.0
google-auth==2.30.0